Work on Build, Test, and Release pipelines

+++ scripts/build-nextjs-site.ps1
@@ -9,21 +9,20 @@ if ($?) {
     exit 1
 }
 
-# Make sure the directory /dist/web-server exists
-if (!(Test-Path -Path "./dist/web-server")) {
-    New-Item -ItemType Directory -Path "./dist/web-server"
+# Create ./dist/web-server-docker-compose
+if (Test-Path -Path "./dist/web-server-docker-compose") {
+    Remove-Item -Path "./dist/web-server-docker-compose" -Recurse -Force
 }
+New-Item -ItemType Directory -Path "./dist/web-server-docker-compose"
 
-Remove-Item -Path "./dist/web-server" -Recurse -Force
-New-Item -ItemType Directory -Path "./dist/web-server"
 
-# Copy the supabase_prod files
-Copy-Item -Path "./supabase_prod/*" -Destination "./dist/web-server" -Recurse -Force
-Copy-Item -Path "./src/db/prod-stuff.sql" -Destination "./dist/web-server/supabase_volumes/db/init/prod-stuff.sql" -Force
+# Copy the docker-compose-setup files
+Copy-Item -Path "./docker-compose-setup/*" -Destination "./dist/web-server-docker-compose" -Recurse -Force
+Copy-Item -Path "./src/db/prod-stuff.sql" -Destination "./dist/web-server-docker-compose/supabase_volumes/db/init/prod-stuff.sql" -Force
 
 # Copy the config files
-New-Item -ItemType Directory -Path "./dist/web-server/config"
-Copy-Item -Path "./config/*" -Destination "./dist/web-server/config/" -Recurse -Force
+New-Item -ItemType Directory -Path "./dist/web-server-docker-compose/config"
+Copy-Item -Path "./config/*" -Destination "./dist/web-server-docker-compose/config/" -Recurse -Force
 
 # get the current version
 $version = (Get-Content -Path "./package.json" | ConvertFrom-Json).version
@@ -45,7 +44,10 @@ if ($?) {
     exit 1
 }
 
-Remove-Item -Path "./dist/web-server-static" -Recurse -Force
+# Create ./dist/web-server-static
+if (Test-Path -Path "./dist/web-server-static") {
+    Remove-Item -Path "./dist/web-server-static" -Recurse -Force
+}
 New-Item -ItemType Directory -Path "./dist/web-server-static"
 
 # Put static files into a directory, ready to be served by a CDN
